Rename .h, .hh, .cc files to .hpp and .cpp
