+++ examples/quickstart/fractals.cpp
@@ -16,7 +16,7 @@
 
 #include <vector>
 
-#include "EasyBMP/EasyBMP.h"
+#include "EasyBMP/EasyBMP.hpp"
 #include "EasyBMP/EasyBMP.cpp"
 
 int const sizeY = 256;
