+++ examples/sheneos/sheneos_compare.cpp
@@ -24,7 +24,7 @@
 #include <vector>
 
 #include "sheneos/interpolator.hpp"
-#include "fname.h"
+#include "fname.hpp"
 
 int omp_thread_count() {
     int n = 0;
