+++ tests/regressions/block_matrix/block_matrix.cpp
@@ -7,9 +7,9 @@
 #if !defined(HPX_COMPUTE_DEVICE_CODE)
 #include <hpx/hpx.hpp>
 
-#include "block_matrix.hh"
+#include "block_matrix.hpp"
 
-#include "matrix.hh"
+#include "matrix.hpp"
 
 #include <hpx/hpx.hpp>
 #include <hpx/include/components.hpp>
