+++ tests/regressions/block_matrix/main.cpp
@@ -7,8 +7,8 @@
 #if !defined(HPX_COMPUTE_DEVICE_CODE)
 #include <hpx/hpx.hpp>
 
-#include "defs.hh"
-#include "tests.hh"
+#include "defs.hpp"
+#include "tests.hpp"
 
 #include <hpx/hpx.hpp>
 #include <hpx/hpx_init.hpp>
