+++ tests/regressions/block_matrix/matrix_hpx.cpp
@@ -7,9 +7,9 @@
 #if !defined(HPX_COMPUTE_DEVICE_CODE)
 #include <hpx/hpx.hpp>
 
-#include "matrix_hpx.hh"
+#include "matrix_hpx.hpp"
 
-#include "algorithms.hh"
+#include "algorithms.hpp"
