+++ tests/regressions/block_matrix/tests.cpp
@@ -7,11 +7,11 @@
 #if !defined(HPX_COMPUTE_DEVICE_CODE)
 #include <hpx/hpx.hpp>
 
-#include "tests.hh"
+#include "tests.hpp"
 
-#include "algorithms.hh"
-#include "block_matrix.hh"
-#include "matrix.hh"
+#include "algorithms.hpp"
+#include "block_matrix.hpp"
+#include "matrix.hpp"
 
 #include <hpx/hpx.hpp>
 #include <hpx/modules/testing.hpp>
